add UI, Parser, and Handler class
