+++ src/main/java/DukeResponse.java
@@ -1,0 +1,19 @@
+
+public class DukeResponse {
+
+    private String response;
+    private int status;
+
+    public DukeResponse(String response, int status) {
+        this.response = response;
+        this.status = status;
+    }
+
+    public String getResponse() {
+        return response;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+}
